+++ lib/pigeon/apns/config.ex
@@ -95,9 +95,17 @@ defmodule Pigeon.APNS.Config do
   @doc false
   def decode_pem(bin) when is_binary(bin) do
     case :public_key.pem_decode(bin) do
-      [{:Certificate, cert, _}] -> cert
-      [{:RSAPrivateKey, key, _}] -> {:RSAPrivateKey, key}
-      _ -> {:error, {:invalid, bin}}
+      [{:Certificate, cert, _}] ->
+        cert
+
+      [{:PrivateKeyInfo, key, _}] ->
+        {:PrivateKeyInfo, key}
+
+      [{:RSAPrivateKey, key, _}] ->
+        {:RSAPrivateKey, key}
+
+      _ ->
+        {:error, {:invalid, bin}}
     end
   end